requirements.txt

huggingface_hub==0.14.1
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
gunicorn